Update ExampleUnitTest.kt
There weren't any unit tests in the project. A lack of unit tests is one of the Developers' Seven Deadly Sins. I understand that there isn't currently enough code to do proper unit tests, but it would still be a good idea to have a baseline for when you are able to do unit tests, since what you have in the unit test files doesn't affect the rest of the project.