Update MainActivity.kt
This particular class has code that may be difficult for some people to figure out at first glance. In these scenarios, it's a good idea to have comments explaining what the code does. Not having enough comments is one of the Developers' Seven Deadly Sins, so it's a good idea to keep an eye out for classes that should have comments.