Update Food.kt
The comment that was written for this class was, at the moment, not necessary. This class wouldn't be difficult for someone to understand when looking at the code, so a comment explaining the class isn't needed. There's another comment in the IExerciseDAO class, but DAO classes may require explanation as they may contain more complex functions. Having too many comments is one of the Developers' Seven Deadly Sins, as is having not enough comments. As such, you want to make sure that it's necessary to put a comment in your code.